feat: Bakery 유효성 검사 구현
