+++ src/main/java/kr/co/breadfeetserver/application/bakery/BakeryService.java
@@ -4,8 +4,8 @@ import kr.co.breadfeetserver.domain.bakery.Bakery;
 import kr.co.breadfeetserver.domain.bakery.BakeryJpaRepository;
 import kr.co.breadfeetserver.infra.exception.BreadFeetBusinessException;
 import kr.co.breadfeetserver.infra.exception.ErrorCode;
-import kr.co.breadfeetserver.presentation.bakery.dto.BakeryCreateRequest;
-import kr.co.breadfeetserver.presentation.bakery.dto.BakeryUpdateRequest;
+import kr.co.breadfeetserver.presentation.bakery.dto.request.BakeryCreateRequest;
+import kr.co.breadfeetserver.presentation.bakery.dto.request.BakeryUpdateRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
