+++ src/main/java/kr/co/breadfeetserver/domain/bakery/Menu.java
@@ -6,7 +6,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import kr.co.breadfeetserver.global.BaseTimeEntity;
+import kr.co.breadfeetserver.global.base.BaseTimeEntity;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
