--- src/main/java/kr/co/breadfeetserver/presentation/bakery/dto/AddressUpdateRequest.java
@@ -1,18 +1,0 @@
-package kr.co.breadfeetserver.presentation.bakery.dto;
-
-import kr.co.breadfeetserver.domain.bakery.AddressJpaVO;
-
-public record AddressUpdateRequest(
-        String detail,
-        String lotNumber,
-        String roadAddress
-) {
-
-    public AddressJpaVO toEntity() {
-        return AddressJpaVO.builder()
-                .detail(detail)
-                .lotNumber(lotNumber)
-                .roadAddress(roadAddress)
-                .build();
-    }
-}
