+++ src/main/java/kr/co/breadfeetserver/presentation/bakery/dto/request/BakeryUpdateRequest.java
@@ -1,12 +1,14 @@
-package kr.co.breadfeetserver.presentation.bakery.dto;
+package kr.co.breadfeetserver.presentation.bakery.dto.request;
 
 import kr.co.breadfeetserver.domain.bakery.Bakery;
+import kr.co.breadfeetserver.global.annotation.IsEssential;
+import kr.co.breadfeetserver.global.annotation.PhoneNumberPattern;
 
 public record BakeryUpdateRequest(
         long bakeryId,
-        String name,
+        @IsEssential String name,
         AddressUpdateRequest address,
-        String phoneNumber,
+        @PhoneNumberPattern String phoneNumber,
         String businessHours,
         String bestBread,
         Double xCoordinate,
