+++ src/test/java/kr/co/breadfeetserver/application/bakery/BakeryServiceTest.java
@@ -11,8 +11,8 @@ import static org.mockito.Mockito.verify;
 import java.util.Optional;
 import kr.co.breadfeetserver.domain.bakery.Bakery;
 import kr.co.breadfeetserver.domain.bakery.BakeryJpaRepository;
-import kr.co.breadfeetserver.presentation.bakery.dto.BakeryCreateRequest;
-import kr.co.breadfeetserver.presentation.bakery.dto.BakeryUpdateRequest;
+import kr.co.breadfeetserver.presentation.bakery.dto.request.BakeryCreateRequest;
+import kr.co.breadfeetserver.presentation.bakery.dto.request.BakeryUpdateRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
