+++ src/test/java/kr/co/breadfeetserver/presentation/bakery/BakeryControllerTest.java
@@ -15,8 +15,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import kr.co.breadfeetserver.application.bakery.BakeryService;
 import kr.co.breadfeetserver.fixture.BakeryFixture;
 import kr.co.breadfeetserver.infra.exception.GlobalExceptionHandler;
-import kr.co.breadfeetserver.presentation.bakery.dto.BakeryCreateRequest;
-import kr.co.breadfeetserver.presentation.bakery.dto.BakeryUpdateRequest;
+import kr.co.breadfeetserver.presentation.bakery.dto.request.BakeryCreateRequest;
+import kr.co.breadfeetserver.presentation.bakery.dto.request.BakeryUpdateRequest;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
